+++ packages/brick_build/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- Add documentation to increase pub.dev score
+
 ## 3.2.1
 
 - Revert `.getDisplayString()` change due to Flutter 3.22 being restricted to analyzer <6.4.1. `meta` is pinned to `1.12` in this version of Flutter, and `analyzer >=6.5.0`, where the change was made, requires `meta >= 1.15`. This change will eventually be re-reverted.