+++ packages/brick_offline_first_with_supabase_build/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- Add documentation to increase pub.dev score
+
 ## 1.0.0
 
 - Stable release